finished functionality of routes folder proceeding to controllers folder
